+++ mmssystem/src/main/java/com/g174/mmssystem/service/Impl/PurchaseRequisitionServiceImpl.java
@@ -133,7 +133,7 @@ public class PurchaseRequisitionServiceImpl implements IPurchaseRequisitionServi
     }
 
     @Override
-    public PurchaseRequisitionResponseDTO getRequisitionById(Long requisitionId) {
+    public PurchaseRequisitionResponseDTO getRequisitionById(Integer requisitionId) {
         log.info("Lấy purchase requisition ID: {}", requisitionId);
 
         PurchaseRequisition requisition = requisitionRepository.findByIdWithRelations(requisitionId)
@@ -188,7 +188,7 @@ public class PurchaseRequisitionServiceImpl implements IPurchaseRequisitionServi
 
     @Override
     @Transactional
-    public PurchaseRequisitionResponseDTO updateRequisition(Long requisitionId, PurchaseRequisitionRequestDTO dto, Integer updatedById) {
+    public PurchaseRequisitionResponseDTO updateRequisition(Integer requisitionId, PurchaseRequisitionRequestDTO dto, Integer updatedById) {
         log.info("Cập nhật purchase requisition ID: {}", requisitionId);
 
         PurchaseRequisition requisition = requisitionRepository.findById(requisitionId)
@@ -292,7 +292,7 @@ public class PurchaseRequisitionServiceImpl implements IPurchaseRequisitionServi
 
     @Override
     @Transactional
-    public PurchaseRequisitionResponseDTO approveRequisition(Long requisitionId, Integer approverId) {
+    public PurchaseRequisitionResponseDTO approveRequisition(Integer requisitionId, Integer approverId) {
         log.info("Approve purchase requisition ID: {} bởi approver ID: {}", requisitionId, approverId);
 
         PurchaseRequisition requisition = requisitionRepository.findById(requisitionId)
@@ -324,7 +324,7 @@ public class PurchaseRequisitionServiceImpl implements IPurchaseRequisitionServi
 
     @Override
     @Transactional
-    public PurchaseRequisitionResponseDTO rejectRequisition(Long requisitionId, Integer approverId, String reason) {
+    public PurchaseRequisitionResponseDTO rejectRequisition(Integer requisitionId, Integer approverId, String reason) {
         log.info("Reject purchase requisition ID: {} bởi approver ID: {} với lý do: {}", requisitionId, approverId, reason);
 
         PurchaseRequisition requisition = requisitionRepository.findById(requisitionId)
@@ -356,7 +356,7 @@ public class PurchaseRequisitionServiceImpl implements IPurchaseRequisitionServi
 
     @Override
     @Transactional
-    public PurchaseRequisitionResponseDTO closeRequisition(Long requisitionId) {
+    public PurchaseRequisitionResponseDTO closeRequisition(Integer requisitionId) {
         log.info("Đóng purchase requisition ID: {}", requisitionId);
 
         PurchaseRequisition requisition = requisitionRepository.findById(requisitionId)
@@ -383,7 +383,7 @@ public class PurchaseRequisitionServiceImpl implements IPurchaseRequisitionServi
 
     @Override
     @Transactional
-    public PurchaseRequisitionResponseDTO restoreRequisition(Long requisitionId) {
+    public PurchaseRequisitionResponseDTO restoreRequisition(Integer requisitionId) {
         log.info("Khôi phục purchase requisition ID: {}", requisitionId);
 
         PurchaseRequisition requisition = requisitionRepository.findById(requisitionId)
@@ -408,7 +408,7 @@ public class PurchaseRequisitionServiceImpl implements IPurchaseRequisitionServi
 
     @Override
     @Transactional
-    public PurchaseRequisitionResponseDTO deleteRequisition(Long requisitionId) {
+    public PurchaseRequisitionResponseDTO deleteRequisition(Integer requisitionId) {
         log.info("Xóa purchase requisition ID: {}", requisitionId);
 
         PurchaseRequisition requisition = requisitionRepository.findById(requisitionId)
